Refactored `usm64` to new `gen` package API
Now, each instruction does not store the information about it's arguments and targets, but instead stores a reference to the relevant `InstructionInfo` instance, and queries it when needed.

This is a less "robust" solution, since we now do lots of validations in emulation time (we interpret the argument types, etc, and do not assume they are already validated), but this allows for future optimization steps to change our arguments and targets "beneath our legs".

Emulation now is just for testing purposes, and is not a "production" feature that we should keep. For that reason only I allow the current quality and robustness of the emulation process (which is very poor).