Added custom control flow graph generation
The `gen` package now create the control flow graph and basic blocks structure without using the `graph` package. It utilizes the direct provided information, and is more efficient and generates a new basic block on any branch, even if it can be optimized to be a single block. This is done to better represent the source code in our internal structures, and to allow the end user to chose it's optimization, and not be forced to optimize this way (It previously could "accidentally" changed the arrangement of basic blocks)